frontend-plugin-api: extract Expand to top-level types
